app_comparer/requirements.txt
# ========================================
# APP COMPARER - Requirements
# Python 3.11.9+
# ========================================

# Google Sheets API
gspread==6.1.2
google-api-python-client==2.139.0
oauth2client==4.1.3

# Configuration
python-dotenv==1.0.1
